+++ lib/src/controls/surfaces/dialog.dart
@@ -303,10 +303,6 @@ class FluentDialogRoute<T> extends RawDialogRoute<T> {
           pageBuilder: (BuildContext context, animation, secondaryAnimation) {
             final pageChild = Builder(builder: builder);
             final dialog = themes?.wrap(pageChild) ?? pageChild;
-
-            /// TODO: CallbackShortcuts + FocusScope is the current workaround for
-            /// <https://github.com/flutter/flutter/issues/97581>, while it's not
-            /// fixed
             return SafeArea(
               child: Actions(
                 actions: {